Move HTTPClientSpy to a file
