+++ CryptoMarketPrice/CryptoMarketPriceTests/LoadMarketListFromRemoteUseCaseTests.swift
@@ -28,25 +28,6 @@ private extension Array where Element == RemoteCryptoMarket {
     }
 }
 
-final class HTTPClientSpy: HTTPClient {
-    var requestURLs = [URL]()
-    var requestCompletions = [(Result) -> Void]()
-    
-    func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) {
-        requestURLs.append(url)
-        requestCompletions.append(completion)
-    }
-    
-    func complete(with error: Error, at index: Int = 0) {
-        requestCompletions[index](.failure(error))
-    }
-    
-    func complete(with statusCode: Int, data: Data, at index: Int = 0) {
-        let response = HTTPURLResponse(url: requestURLs[index], statusCode: statusCode, httpVersion: nil, headerFields: nil)!
-        requestCompletions[index](.success((data, response)))
-    }
-}
-
 final class RemoteCryptoMarketLoader: CryptoMarketLoader {
     enum LoadError: Swift.Error {
         case invalidData
